+++ pages/admin/main.tsx
@@ -1,16 +1,59 @@
+/* eslint-disable react-hooks/rules-of-hooks */
 import { NextPage } from 'next';
 import Head from 'next/head';
-import { AdminMainTemplete } from 'public';
+import { AdminMainTemplete, TaskAction } from 'public';
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import { io } from 'socket.io-client';
 
 const adminMainPage: NextPage = () => {
+  const socket = io();
+  const taskAction = new TaskAction();
+  const dispatch = useDispatch();
+
+  const socketFlug = () => {
+    console.log('通信きた');
+    fetch(`${process.env.NEXT_PUBLIC_URL}/getAdminTask`)
+      .then((res) => {
+        return res.json();
+      })
+      .then((json) => dispatch(taskAction.setAdminTasks(json)));
+  };
+
+  useEffect(() => {
+    //Serverからメッセージを受信
+    socket.on('crResult', (data: { message: boolean }) => {
+      if (data.message) {
+        socketFlug();
+        dispatch(taskAction.setPage('list'));
+      } else {
+        console.log('error');
+      }
+    });
+
+    fetch(`${process.env.NEXT_PUBLIC_URL}/getAdminTask`)
+      .then((res) => {
+        return res.json();
+      })
+      .then((json) => {
+        dispatch(taskAction.setAdminTasks(json));
+      });
+  }, []);
+
   return (
     <>
       <Head>
         <title>Admin</title>
       </Head>
-      <AdminMainTemplete />
+      <AdminMainTemplete socket={socket} />
     </>
   );
 };
+
+// すべてのリクエストの度に実行される;
+export function getServerSideProps() {
+  // props を通じて Page に data を渡す
+  return { props: {} };
+}
 
 export default adminMainPage;
