+++ styles/theme.ts
@@ -6,24 +6,26 @@ const theme = extendTheme({
       200: '#0B1C9D',
       300: '#1b85d1',
     },
-    LimeGreen: {
+    marineProduct: {
+      100: '#2599EF', //下に行くほど暗い色
+      200: '#0B1C9D',
+      300: '#1b85d1',
+    },
+    agriculturalProduct: {
       100: '#7CE032',
       200: '#43B211',
     },
-    vividRed: {
+    livestock: {
       100: '#FF7983',
       200: '#AC0922',
     },
-    cream: {
-      100: '#fffaf0',
-    },
-    vividYellow: {
+    delicatessen: {
       100: '#ffd700',
     },
-    thinPink: {
+    serviceCounter: {
       100: '#ee82ee',
     },
-    thinRightBlue: {
+    glossary: {
       100: '#00ffff',
     },
   },
